+++ examples/GD32F303_FOCExplore/software/Application/Include/gd32f30x_it.h
@@ -14,7 +14,7 @@
 #include "LED.h"
 #include "usart0.h"
 #include "usart2.h"
-#include "timer1.h"
+#include "timer5.h"
 #include "timer2.h"
 #include "auxtimer.h"
 #include "adc.h"
@@ -47,8 +47,8 @@ void TIMER0_UP_IRQHandler(void);
 void USART0_IRQHandler(void);
 /* this function handles USART2 interrupt */
 void USART2_IRQHandler(void);
-/* this function handles TIMER1 interrupt */
-void TIMER1_IRQHandler(void);
+/* this function handles TIMER5 interrupt (control scheduler tick) */
+void TIMER5_IRQHandler(void);
 /* this function handles TIMER2 interrupt */
 void TIMER2_IRQHandler(void);
 /* this function handles TIMER4 interrupt */